Adding an option to create and choose ESP for UEFI install
The installer already had code to check for ESP on UEFI install.
However, this automatic check is pretty limited, not to mention on
a blank disk like a VM it doesn't even tell user to create an ESP.
So I decided to create a whole new ESP creating and choosing stage
before install. Most of the code was based on the current code to choose
and create installation drive. Some part of the install stage can be able
to split and use with the ESP stage so I take advantage of that too.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>