Create a script to generate iso image from target zip files
To prepare for producing signed releases in an iso image, we have to
prepare a script to convert target zips (that are signed) into iso
image. This script need to get input of the zip, where is newinstaller
and install.img localtion so it will be able to produce the iso as
output.

Also, to be able to build install.img standalone, we need to remove
efibootmgr and put them straight into the system.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>